+++ cstar/orchestration/formatting.py
@@ -1,7 +1,5 @@
 import typing as t
 
-from cstar.orchestration.orchestration import RunRequest
-
 TFormattable = t.TypeVar("TFormattable", contravariant=True)
 
 
@@ -45,26 +43,3 @@ class ModelFormatter(t.Protocol, t.Generic[TFormattable]):
         if updates:
             s = self._apply_replacements(s, updates)
         return s
-
-
-class RunRequestCommandFormatter(ModelFormatter[RunRequest]):
-    """Format a `RunRequest` as a request as a CLI command."""
-
-    def _to_string(self, value: RunRequest) -> str:
-        variables = " ".join(f"{k}='{v}'" for k, v in value.environment.items())
-        cmd = " ".join(value.command)
-
-        return f"{variables} {cmd}".strip()
-
-
-class RunRequestScriptFormatter(ModelFormatter[RunRequest]):
-    """Format a `RunRequest` as script content."""
-
-    def _to_string(self, value: RunRequest) -> str:
-        command = " ".join(value.command)
-        exports = ";".join(f"export {k}='{v}'" for k, v in value.environment.items())
-
-        if exports:
-            return f"{exports}; {command};"
-
-        return f"{command};"
